Moved sidebar from templates to components
